+++ src/renderer/src/scripts/metadata-files/datasetDescription.js
@@ -416,7 +416,7 @@ window.generateDDFile = async (uploadBFBoolean) => {
               This dataset is currently being reviewed by the SPARC curation team, therefore, has been set to read-only mode. No changes can be made to this dataset until the review is complete.
               <br />
               <br />
-              If you would like to make changes to this dataset, please reach out to the SPARC curation team at <a href="mailto:curation@sparc.science" target="_blank">curation@sparc.science.</a>
+              If you would like to make changes to this dataset, please reach out to the SPARC curation team at <a  target="_blank" href="mailto:curation@sparc.science">curation@sparc.science.</a>
             `,
         width: 600,
         heightAuto: false,