+++ src/renderer/src/scripts/others/announcements.js
@@ -1,5 +1,7 @@
 import axios from "axios";
 import Swal from "sweetalert2";
+import lottie from "lottie-web";
+import { announcement_laptop } from "../../assets/lotties/announcement-laptop";
 
 // To change what branch the announcements.json is fetched from, enter branch name between "SODA-for-SPARC/" and "/scripts" in the url below
 // state will be either "update" or "announcements"
@@ -36,26 +38,26 @@ const checkForAnnouncements = async (state) => {
 
       if (features && features.length > 0) {
         htmlMessageFeatures = `
-              <label style="font-weight: 700; font-size: 17px;">Feature Additions:<br></label>
+              <label style="font-weight: 700; font-size: 17px;">Feature Additions:<br></label><ul>
               ${features
                 .map((feature) => {
                   return `<li style="margin: .5rem 0 .5rem 0;">${feature}</li>`;
                 })
                 .join("")}
-        `;
+                </ul>`;
       }
 
       let htmlMessageBugFixes = ``;
 
       if (bugFixes && bugFixes.length > 0) {
         htmlMessageBugFixes = `
-              <label style="font-weight: 700; font-size: 17px;">Bug Fixes:<br></label>
+              <label style="font-weight: 700; font-size: 17px;">Bug Fixes:<br></label><ul>
               ${bugFixes
                 .map((bugfix) => {
                   return `<li style="margin: .5rem 0 .5rem 0;">${bugfix}</li>`;
                 })
                 .join("")}
-        `;
+                </ul>`;
       }
       let htmlMessage = `
           <div style="text-align: justify; overflow-y: auto; max-height: 350px;">
@@ -67,17 +69,45 @@ const checkForAnnouncements = async (state) => {
           `;
 
       await Swal.fire({
-        title: `Welcome to SODA for SPARC ${appVersion} for ${platform}`,
+        title: `What's new in SODA V${appVersion}`,
         html: htmlMessage,
         icon: "info",
         heightAuto: false,
         backdrop: "rgba(0,0,0, 0.4)",
-        confirmButtonText: "Okay",
+        confirmButtonText: "Get Started",
         allowOutsideClick: false,
         allowEscapeKey: false,
         didOpen: () => {
           let swal_alert = document.getElementsByClassName("swal2-popup")[0];
           swal_alert.style.width = "60rem";
+          console.log("did open");
+
+          let html_container = document.getElementById("swal2-html-container");
+          let html_child = html_container.children[0];
+          html_child.style.border = "2px solid #13716d";
+          html_child.style.padding = "1rem";
+          html_container.style.marginTop = "0";
+          // Create lottie animation for the title
+          let swal_icon = document.getElementsByClassName("swal2-icon")[0];
+          swal_icon.innerHTML = "";
+          swal_icon.style.border = "none";
+          swal_icon.style.width = "10rem";
+          swal_icon.style.height = "9rem";
+          swal_icon.style.marginTop = "0";
+          swal_icon.style.marginBottom = "0";
+
+          lottie.loadAnimation({
+            rendererSettings: {
+              preserveAspectRatio: "xMidYMid meet",
+            },
+            width: 400,
+            height: 400,
+            container: swal_icon,
+            animationData: announcement_laptop,
+            renderer: "svg",
+            loop: true,
+            autoplay: true,
+          });
         },
       });
     } else if (state === "update") {